requirements.txt
async-timeout==5.0.1
blinker==1.9.0
click==8.1.8
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
redis==6.0.0
SQLAlchemy==2.0.40
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
gunicorn==22.0.0